mano/requirements.txt
requests==2.20.1
Flask==1.0.2
gevent==1.4.0
pymongo==3.8.0
Flask-RESTful==0.3.7
PyYAML
Flask-Cors==3.0.8
openstacksdk==0.38.0
configobj==5.0.6
jsonschema==2.6.0